refactored to rest controller
