+++ server/src/main/java/com/omigost/server/config/AWSLocalstackConfig.java
@@ -3,6 +3,7 @@ package com.omigost.server.config;
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.auth.AWSCredentialsProvider;
 import com.amazonaws.services.budgets.AWSBudgets;
 import com.amazonaws.services.budgets.AWSBudgetsClientBuilder;
 import com.amazonaws.services.costexplorer.AWSCostExplorer;
@@ -80,7 +81,9 @@ public class AWSLocalstackConfig {
 
     //TODO get iam credentials for root
     @Bean(name = "iamCredentials")
-    AWSCredentialsProvider iamRootCredentials() {return  awsContainer.getCredentialsProvider();}
+    AWSCredentialsProvider iamRootCredentials() {
+        return awsContainer.getCredentialsProvider();
+    }
 
     @Bean
     public AmazonIdentityManagement amazonIdentityManagement() {
