+++ server/src/main/java/com/omigost/server/notification/NotificationService.java
@@ -4,4 +4,5 @@ import com.omigost.server.model.Communication;
 
 public interface NotificationService {
     void sendAlertToUser(Communication communication, NotificationMessage message);
+
 }
